pyproject.toml
[project]
name = "jp-holidays"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]
